compatibility isolation pass 3
