+++ .gitignore
@@ -42,3 +42,4 @@
 /PlayerHeads-common-support/target/
 /PlayerHeads-compatibility-lib/target/
 /PlayerHeads-common-support/nbproject/
+/PlayerHeads-base-api/target/
