+++ PlayerHeads-api/docs/allclasses-frame.html
@@ -12,13 +12,7 @@
 <h1 class="bar">All&nbsp;Classes</h1>
 <div class="indexContainer">
 <ul>
-<li><a href="org/shininet/bukkit/playerheads/events/BlockDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">BlockDropHeadEvent</a></li>
-<li><a href="org/shininet/bukkit/playerheads/events/DropHeadEvent.html" title="interface in org.shininet.bukkit.playerheads.events" target="classFrame"><span class="interfaceName">DropHeadEvent</span></a></li>
-<li><a href="org/shininet/bukkit/playerheads/events/HeadRollEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">HeadRollEvent</a></li>
 <li><a href="com/github/crashdemons/playerheads/api/HeadType.html" title="interface in com.github.crashdemons.playerheads.api" target="classFrame"><span class="interfaceName">HeadType</span></a></li>
-<li><a href="org/shininet/bukkit/playerheads/events/LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">LivingEntityDropHeadEvent</a></li>
-<li><a href="org/shininet/bukkit/playerheads/events/MobDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">MobDropHeadEvent</a></li>
-<li><a href="org/shininet/bukkit/playerheads/events/PlayerDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">PlayerDropHeadEvent</a></li>
 <li><a href="com/github/crashdemons/playerheads/api/PlayerHeads.html" title="class in com.github.crashdemons.playerheads.api" target="classFrame">PlayerHeads</a></li>
 <li><a href="com/github/crashdemons/playerheads/api/PlayerHeadsAPI.html" title="interface in com.github.crashdemons.playerheads.api" target="classFrame"><span class="interfaceName">PlayerHeadsAPI</span></a></li>
 <li><a href="org/shininet/bukkit/playerheads/PlayerHeadsPlugin.html" title="interface in org.shininet.bukkit.playerheads" target="classFrame"><span class="interfaceName">PlayerHeadsPlugin</span></a></li>
