+++ PlayerHeads-api/docs/deprecated-list.html
@@ -74,7 +74,6 @@
 <h2 title="Contents">Contents</h2>
 <ul>
 <li><a href="#method">Deprecated Methods</a></li>
-<li><a href="#constructor">Deprecated Constructors</a></li>
 </ul>
 </div>
 <div class="contentContainer"><a name="method">
@@ -102,27 +101,6 @@
 <td class="colOne"><a href="com/github/crashdemons/playerheads/api/PlayerHeads.html#setApiInstance-com.github.crashdemons.playerheads.api.PlayerHeadsAPI-">com.github.crashdemons.playerheads.api.PlayerHeads.setApiInstance(PlayerHeadsAPI)</a>
 <div class="block"><span class="deprecationComment">This method is intended only for internal PlayerHeads use.</span></div>
 </td>
-</tr>
-</tbody>
-</table>
-</li>
-</ul>
-<a name="constructor">
-<!--   -->
-</a>
-<ul class="blockList">
-<li class="blockList">
-<table class="deprecatedSummary" border="0" cellpadding="3" cellspacing="0" summary="Deprecated Constructors table, listing deprecated constructors, and an explanation">
-<caption><span>Deprecated Constructors</span><span class="tabEnd">&nbsp;</span></caption>
-<tr>
-<th class="colOne" scope="col">Constructor and Description</th>
-</tr>
-<tbody>
-<tr class="altColor">
-<td class="colOne"><a href="org/shininet/bukkit/playerheads/events/HeadRollEvent.html#HeadRollEvent-org.bukkit.entity.Entity-org.bukkit.entity.Entity-boolean-double-double-double-double-double-boolean-">org.shininet.bukkit.playerheads.events.HeadRollEvent(Entity, Entity, boolean, double, double, double, double, double, boolean)</a></td>
-</tr>
-<tr class="rowColor">
-<td class="colOne"><a href="org/shininet/bukkit/playerheads/events/HeadRollEvent.html#HeadRollEvent-org.bukkit.entity.Entity-org.bukkit.entity.Entity-boolean-double-double-double-double-double-double-boolean-">org.shininet.bukkit.playerheads.events.HeadRollEvent(Entity, Entity, boolean, double, double, double, double, double, double, boolean)</a></td>
 </tr>
 </tbody>
 </table>
