--- PlayerHeads-api/docs/org/shininet/bukkit/playerheads/events/class-use/LivingEntityDropHeadEvent.html
@@ -1,175 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc -->
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>Uses of Class org.shininet.bukkit.playerheads.events.LivingEntityDropHeadEvent (PlayerHeads-api 5.2.11-SNAPSHOT API)</title>
-<link rel="stylesheet" type="text/css" href="../../../../../../stylesheet.css" title="Style">
-<script type="text/javascript" src="../../../../../../script.js"></script>
-</head>
-<body>
-<script type="text/javascript"><!--
-    try {
-        if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="Uses of Class org.shininet.bukkit.playerheads.events.LivingEntityDropHeadEvent (PlayerHeads-api 5.2.11-SNAPSHOT API)";
-        }
-    }
-    catch(err) {
-    }
-//-->
-</script>
-<noscript>
-<div>JavaScript is disabled on your browser.</div>
-</noscript>
-<!-- ========= START OF TOP NAVBAR ======= -->
-<div class="topNav"><a name="navbar.top">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.top" title="Skip navigation links">Skip navigation links</a></div>
-<a name="navbar.top.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../../../../../../overview-summary.html">Overview</a></li>
-<li><a href="../package-summary.html">Package</a></li>
-<li><a href="../../../../../../org/shininet/bukkit/playerheads/events/LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events">Class</a></li>
-<li class="navBarCell1Rev">Use</li>
-<li><a href="../../../../../../overview-tree.html">Tree</a></li>
-<li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
-<li><a href="../../../../../../index-all.html">Index</a></li>
-<li><a href="../../../../../../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li>Prev</li>
-<li>Next</li>
-</ul>
-<ul class="navList">
-<li><a href="../../../../../../index.html?org/shininet/bukkit/playerheads/events/class-use/LivingEntityDropHeadEvent.html" target="_top">Frames</a></li>
-<li><a href="LivingEntityDropHeadEvent.html" target="_top">No&nbsp;Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_top">
-<li><a href="../../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_top");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip.navbar.top">
-<!--   -->
-</a></div>
-<!-- ========= END OF TOP NAVBAR ========= -->
-<div class="header">
-<h2 title="Uses of Class org.shininet.bukkit.playerheads.events.LivingEntityDropHeadEvent" class="title">Uses of Class<br>org.shininet.bukkit.playerheads.events.LivingEntityDropHeadEvent</h2>
-</div>
-<div class="classUseContainer">
-<ul class="blockList">
-<li class="blockList">
-<table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing packages, and an explanation">
-<caption><span>Packages that use <a href="../../../../../../org/shininet/bukkit/playerheads/events/LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events">LivingEntityDropHeadEvent</a></span><span class="tabEnd">&nbsp;</span></caption>
-<tr>
-<th class="colFirst" scope="col">Package</th>
-<th class="colLast" scope="col">Description</th>
-</tr>
-<tbody>
-<tr class="altColor">
-<td class="colFirst"><a href="#org.shininet.bukkit.playerheads.events">org.shininet.bukkit.playerheads.events</a></td>
-<td class="colLast">
-<div class="block">Defines events created by the PlayerHeads plugin, including beheadings.</div>
-</td>
-</tr>
-</tbody>
-</table>
-</li>
-<li class="blockList">
-<ul class="blockList">
-<li class="blockList"><a name="org.shininet.bukkit.playerheads.events">
-<!--   -->
-</a>
-<h3>Uses of <a href="../../../../../../org/shininet/bukkit/playerheads/events/LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events">LivingEntityDropHeadEvent</a> in <a href="../../../../../../org/shininet/bukkit/playerheads/events/package-summary.html">org.shininet.bukkit.playerheads.events</a></h3>
-<table class="useSummary" border="0" cellpadding="3" cellspacing="0" summary="Use table, listing subclasses, and an explanation">
-<caption><span>Subclasses of <a href="../../../../../../org/shininet/bukkit/playerheads/events/LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events">LivingEntityDropHeadEvent</a> in <a href="../../../../../../org/shininet/bukkit/playerheads/events/package-summary.html">org.shininet.bukkit.playerheads.events</a></span><span class="tabEnd">&nbsp;</span></caption>
-<tr>
-<th class="colFirst" scope="col">Modifier and Type</th>
-<th class="colLast" scope="col">Class and Description</th>
-</tr>
-<tbody>
-<tr class="altColor">
-<td class="colFirst"><code>class&nbsp;</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../../org/shininet/bukkit/playerheads/events/MobDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events">MobDropHeadEvent</a></span></code>
-<div class="block">Event created by the PlayerHeads plugin when a Mob is beheaded.</div>
-</td>
-</tr>
-<tr class="rowColor">
-<td class="colFirst"><code>class&nbsp;</code></td>
-<td class="colLast"><code><span class="memberNameLink"><a href="../../../../../../org/shininet/bukkit/playerheads/events/PlayerDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events">PlayerDropHeadEvent</a></span></code>
-<div class="block">Event created by the PlayerHeads plugin when a Player is beheaded.</div>
-</td>
-</tr>
-</tbody>
-</table>
-</li>
-</ul>
-</li>
-</ul>
-</div>
-<!-- ======= START OF BOTTOM NAVBAR ====== -->
-<div class="bottomNav"><a name="navbar.bottom">
-<!--   -->
-</a>
-<div class="skipNav"><a href="#skip.navbar.bottom" title="Skip navigation links">Skip navigation links</a></div>
-<a name="navbar.bottom.firstrow">
-<!--   -->
-</a>
-<ul class="navList" title="Navigation">
-<li><a href="../../../../../../overview-summary.html">Overview</a></li>
-<li><a href="../package-summary.html">Package</a></li>
-<li><a href="../../../../../../org/shininet/bukkit/playerheads/events/LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events">Class</a></li>
-<li class="navBarCell1Rev">Use</li>
-<li><a href="../../../../../../overview-tree.html">Tree</a></li>
-<li><a href="../../../../../../deprecated-list.html">Deprecated</a></li>
-<li><a href="../../../../../../index-all.html">Index</a></li>
-<li><a href="../../../../../../help-doc.html">Help</a></li>
-</ul>
-</div>
-<div class="subNav">
-<ul class="navList">
-<li>Prev</li>
-<li>Next</li>
-</ul>
-<ul class="navList">
-<li><a href="../../../../../../index.html?org/shininet/bukkit/playerheads/events/class-use/LivingEntityDropHeadEvent.html" target="_top">Frames</a></li>
-<li><a href="LivingEntityDropHeadEvent.html" target="_top">No&nbsp;Frames</a></li>
-</ul>
-<ul class="navList" id="allclasses_navbar_bottom">
-<li><a href="../../../../../../allclasses-noframe.html">All&nbsp;Classes</a></li>
-</ul>
-<div>
-<script type="text/javascript"><!--
-  allClassesLink = document.getElementById("allclasses_navbar_bottom");
-  if(window==top) {
-    allClassesLink.style.display = "block";
-  }
-  else {
-    allClassesLink.style.display = "none";
-  }
-  //-->
-</script>
-</div>
-<a name="skip.navbar.bottom">
-<!--   -->
-</a></div>
-<!-- ======== END OF BOTTOM NAVBAR ======= -->
-<p class="legalCopy"><small>Copyright &#169; 2020 <a href="http://www.shininet.org">ShiniNet</a>. All rights reserved.</small></p>
-</body>
-</html>
