--- PlayerHeads-api/docs/org/shininet/bukkit/playerheads/events/package-frame.html
@@ -1,28 +1,0 @@
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<!-- NewPage -->
-<html lang="en">
-<head>
-<!-- Generated by javadoc -->
-<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>org.shininet.bukkit.playerheads.events (PlayerHeads-api 5.2.11-SNAPSHOT API)</title>
-<link rel="stylesheet" type="text/css" href="../../../../../stylesheet.css" title="Style">
-<script type="text/javascript" src="../../../../../script.js"></script>
-</head>
-<body>
-<h1 class="bar"><a href="../../../../../org/shininet/bukkit/playerheads/events/package-summary.html" target="classFrame">org.shininet.bukkit.playerheads.events</a></h1>
-<div class="indexContainer">
-<h2 title="Interfaces">Interfaces</h2>
-<ul title="Interfaces">
-<li><a href="DropHeadEvent.html" title="interface in org.shininet.bukkit.playerheads.events" target="classFrame"><span class="interfaceName">DropHeadEvent</span></a></li>
-</ul>
-<h2 title="Classes">Classes</h2>
-<ul title="Classes">
-<li><a href="BlockDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">BlockDropHeadEvent</a></li>
-<li><a href="HeadRollEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">HeadRollEvent</a></li>
-<li><a href="LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">LivingEntityDropHeadEvent</a></li>
-<li><a href="MobDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">MobDropHeadEvent</a></li>
-<li><a href="PlayerDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events" target="classFrame">PlayerDropHeadEvent</a></li>
-</ul>
-</div>
-</body>
-</html>
