+++ PlayerHeads-api/docs/org/shininet/bukkit/playerheads/package-summary.html
@@ -44,7 +44,7 @@
 <div class="subNav">
 <ul class="navList">
 <li><a href="../../../../com/github/crashdemons/playerheads/api/package-summary.html">Prev&nbsp;Package</a></li>
-<li><a href="../../../../org/shininet/bukkit/playerheads/events/package-summary.html">Next&nbsp;Package</a></li>
+<li>Next&nbsp;Package</li>
 </ul>
 <ul class="navList">
 <li><a href="../../../../index.html?org/shininet/bukkit/playerheads/package-summary.html" target="_top">Frames</a></li>
@@ -124,7 +124,7 @@
 <div class="subNav">
 <ul class="navList">
 <li><a href="../../../../com/github/crashdemons/playerheads/api/package-summary.html">Prev&nbsp;Package</a></li>
-<li><a href="../../../../org/shininet/bukkit/playerheads/events/package-summary.html">Next&nbsp;Package</a></li>
+<li>Next&nbsp;Package</li>
 </ul>
 <ul class="navList">
 <li><a href="../../../../index.html?org/shininet/bukkit/playerheads/package-summary.html" target="_top">Frames</a></li>
