+++ PlayerHeads-api/docs/org/shininet/bukkit/playerheads/package-tree.html
@@ -44,7 +44,7 @@
 <div class="subNav">
 <ul class="navList">
 <li><a href="../../../../com/github/crashdemons/playerheads/api/package-tree.html">Prev</a></li>
-<li><a href="../../../../org/shininet/bukkit/playerheads/events/package-tree.html">Next</a></li>
+<li>Next</li>
 </ul>
 <ul class="navList">
 <li><a href="../../../../index.html?org/shininet/bukkit/playerheads/package-tree.html" target="_top">Frames</a></li>
@@ -129,7 +129,7 @@
 <div class="subNav">
 <ul class="navList">
 <li><a href="../../../../com/github/crashdemons/playerheads/api/package-tree.html">Prev</a></li>
-<li><a href="../../../../org/shininet/bukkit/playerheads/events/package-tree.html">Next</a></li>
+<li>Next</li>
 </ul>
 <ul class="navList">
 <li><a href="../../../../index.html?org/shininet/bukkit/playerheads/package-tree.html" target="_top">Frames</a></li>
