+++ PlayerHeads-api/docs/overview-frame.html
@@ -15,7 +15,6 @@
 <ul title="Packages">
 <li><a href="com/github/crashdemons/playerheads/api/package-frame.html" target="packageFrame">com.github.crashdemons.playerheads.api</a></li>
 <li><a href="org/shininet/bukkit/playerheads/package-frame.html" target="packageFrame">org.shininet.bukkit.playerheads</a></li>
-<li><a href="org/shininet/bukkit/playerheads/events/package-frame.html" target="packageFrame">org.shininet.bukkit.playerheads.events</a></li>
 </ul>
 </div>
 <p>&nbsp;</p>
