+++ PlayerHeads-api/docs/overview-summary.html
@@ -92,12 +92,6 @@
 <div class="block">Package defining api-compatible types for the main plugin classes</div>
 </td>
 </tr>
-<tr class="altColor">
-<td class="colFirst"><a href="org/shininet/bukkit/playerheads/events/package-summary.html">org.shininet.bukkit.playerheads.events</a></td>
-<td class="colLast">
-<div class="block">Defines events created by the PlayerHeads plugin, including beheadings.</div>
-</td>
-</tr>
 </tbody>
 </table>
 </div>
