+++ PlayerHeads-api/docs/overview-tree.html
@@ -74,8 +74,7 @@
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="com/github/crashdemons/playerheads/api/package-tree.html">com.github.crashdemons.playerheads.api</a>, </li>
-<li><a href="org/shininet/bukkit/playerheads/package-tree.html">org.shininet.bukkit.playerheads</a>, </li>
-<li><a href="org/shininet/bukkit/playerheads/events/package-tree.html">org.shininet.bukkit.playerheads.events</a></li>
+<li><a href="org/shininet/bukkit/playerheads/package-tree.html">org.shininet.bukkit.playerheads</a></li>
 </ul>
 </div>
 <div class="contentContainer">
@@ -83,26 +82,6 @@
 <ul>
 <li type="circle">java.lang.<a href="https://docs.oracle.com/javase/8/docs/api/java/lang/Object.html?is-external=true" title="class or interface in java.lang"><span class="typeNameLink">Object</span></a>
 <ul>
-<li type="circle">org.bukkit.event.Event
-<ul>
-<li type="circle">org.bukkit.event.block.BlockEvent
-<ul>
-<li type="circle">org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/BlockDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events"><span class="typeNameLink">BlockDropHeadEvent</span></a> (implements org.bukkit.event.Cancellable, org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/DropHeadEvent.html" title="interface in org.shininet.bukkit.playerheads.events">DropHeadEvent</a>)</li>
-</ul>
-</li>
-<li type="circle">org.bukkit.event.entity.EntityEvent
-<ul>
-<li type="circle">org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/LivingEntityDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events"><span class="typeNameLink">LivingEntityDropHeadEvent</span></a> (implements org.bukkit.event.Cancellable, org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/DropHeadEvent.html" title="interface in org.shininet.bukkit.playerheads.events">DropHeadEvent</a>)
-<ul>
-<li type="circle">org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/MobDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events"><span class="typeNameLink">MobDropHeadEvent</span></a></li>
-<li type="circle">org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/PlayerDropHeadEvent.html" title="class in org.shininet.bukkit.playerheads.events"><span class="typeNameLink">PlayerDropHeadEvent</span></a></li>
-</ul>
-</li>
-</ul>
-</li>
-<li type="circle">org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/HeadRollEvent.html" title="class in org.shininet.bukkit.playerheads.events"><span class="typeNameLink">HeadRollEvent</span></a></li>
-</ul>
-</li>
 <li type="circle">com.github.crashdemons.playerheads.api.<a href="com/github/crashdemons/playerheads/api/PlayerHeads.html" title="class in com.github.crashdemons.playerheads.api"><span class="typeNameLink">PlayerHeads</span></a></li>
 </ul>
 </li>
@@ -129,7 +108,6 @@
 </li>
 </ul>
 </li>
-<li type="circle">org.shininet.bukkit.playerheads.events.<a href="org/shininet/bukkit/playerheads/events/DropHeadEvent.html" title="interface in org.shininet.bukkit.playerheads.events"><span class="typeNameLink">DropHeadEvent</span></a></li>
 <li type="circle">com.github.crashdemons.playerheads.api.<a href="com/github/crashdemons/playerheads/api/HeadType.html" title="interface in com.github.crashdemons.playerheads.api"><span class="typeNameLink">HeadType</span></a></li>
 <li type="circle">com.github.crashdemons.playerheads.api.<a href="com/github/crashdemons/playerheads/api/PlayerHeadsAPI.html" title="interface in com.github.crashdemons.playerheads.api"><span class="typeNameLink">PlayerHeadsAPI</span></a></li>
 <li type="circle">org.bukkit.command.TabCompleter
