+++ PlayerHeads-api/docs/package-list
@@ -1,3 +1,2 @@
 com.github.crashdemons.playerheads.api
 org.shininet.bukkit.playerheads
-org.shininet.bukkit.playerheads.events
