+++ PlayerHeads-api/pom.xml
@@ -35,6 +35,11 @@
         </dependency>
         <dependency>
             <groupId>org.shininet.bukkit</groupId>
+            <artifactId>PlayerHeads-base-api</artifactId>
+            <version>5.2.11-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>org.shininet.bukkit</groupId>
             <artifactId>PlayerHeads-compatibility-api</artifactId>
             <version>5.2.11-SNAPSHOT</version>
         </dependency>
@@ -83,6 +88,28 @@
                     </execution>
                 </executions>
             </plugin>
+
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-shade-plugin</artifactId>
+                <version>3.2.0</version>
+                <executions>
+                    <execution>
+                        <phase>package</phase>
+                        <goals>
+                            <goal>shade</goal>
+                        </goals>
+                        <configuration>
+                            <artifactSet>
+                                <includes>
+                                    <include>org.shininet.bukkit:PlayerHeads-base-api:jar:</include>
+                                </includes>
+                            </artifactSet>
+                        </configuration>
+                    </execution>
+                </executions>
+            </plugin>
+            
             <plugin>
                 <artifactId>maven-jar-plugin</artifactId>
                 <version>3.0.2</version>
