+++ PlayerHeads-base-api/src/main/java/org/shininet/bukkit/playerheads/events/package-info.java
