+++ PlayerHeads-common-support/pom.xml
@@ -125,53 +125,6 @@
                     </archive>
                 </configuration>
             </plugin>
-            <plugin>
-                <artifactId>maven-antrun-plugin</artifactId>
-                <version>1.8</version>
-                <executions>
-                    <execution>
-                        <phase>package</phase>
-                        <configuration>
-                            <target>
-                                <copy file="${project.build.directory}/${project.build.finalName}.${project.packaging}"
-                                      tofile="${project.build.directory}/${project.name}.${project.packaging}"/>
-                                
-                                <copy file="${project.build.directory}/${project.build.finalName}.${project.packaging}"
-                                      tofile="${basedir}/../target/${project.build.finalName}.${project.packaging}"/>
-                                <copy file="${project.build.directory}/${project.name}.${project.packaging}"
-                                      tofile="${basedir}/../target/${project.name}.${project.packaging}"/>
-                                <!--
-                                <copy file="${project.build.directory}/${project.build.finalName}-javadoc.${project.packaging}"
-                                      tofile="${basedir}/../target/${project.build.finalName}-javadoc.${project.packaging}"/>
-                                <copy file="${project.build.directory}/${project.build.finalName}-javadoc.${project.packaging}"
-                                      tofile="${basedir}/../target/${project.name}-javadoc.${project.packaging}"/>
-                                                                            -->
-                            </target>
-                        </configuration>
-                        <goals>
-                            <goal>run</goal>
-                        </goals>
-                    </execution>
-                    <!--
-                    <execution>
-                        <id>copy-repo-docs</id>
-                        <phase>install</phase>
-                        <configuration>
-                            <target>
-                                <delete includeEmptyDirs="true" failonerror="false">
-                                    <fileset dir="${basedir}/docs" includes="**/*" />
-                                </delete>
-                                <copy todir="${basedir}/docs">
-                                    <fileset dir="${project.build.directory}/apidocs" includes="**/*" />
-                                </copy>
-                            </target>
-                        </configuration>
-                        <goals>
-                            <goal>run</goal>
-                        </goals>
-                    </execution>-->
-                </executions>
-            </plugin>
         </plugins>
     </build>
     <!--
