+++ PlayerHeads-compatibility-lib/src/test/java/com/github/crashdemons/playerheads/compatibility/CompatibilitySupportTest.java
