--- PlayerHeads-core/src/main/java/com/github/crashdemons/playerheads/compatibility/plugins/heads/HeadModificationHandling.java
@@ -1,15 +1,0 @@
-/*
- *  This Source Code Form is subject to the terms of the Mozilla Public
- *  License, v. 2.0. If a copy of the MPL was not distributed with this
- *  file, You can obtain one at http://mozilla.org/MPL/2.0/ .
- */
-package com.github.crashdemons.playerheads.compatibility.plugins.heads;
-
-/**
- *
- * @author crashdemons (crashenator at gmail.com)
- */
-public enum HeadModificationHandling {
-    NORMAL,
-    NO_INTERACTION,
-}
