+++ PlayerHeads-crossversion-plugin/dependency-reduced-pom.xml
@@ -3,7 +3,7 @@
   <parent>
     <artifactId>PlayerHeads</artifactId>
     <groupId>org.shininet.bukkit</groupId>
-    <version>5.2.10-SNAPSHOT</version>
+    <version>5.2.11-SNAPSHOT</version>
   </parent>
   <modelVersion>4.0.0</modelVersion>
   <groupId>org.shininet.bukkit</groupId>
@@ -149,21 +149,22 @@
             <configuration>
               <artifactSet>
                 <includes>
+                  <include>org.shininet.bukkit:PlayerHeads-compatibility-lib:jar:</include>
                   <include>org.shininet.bukkit:PlayerHeads-api:jar:</include>
                   <include>org.shininet.bukkit:PlayerHeads-core:jar:</include>
-                  <include>org.shininet.bukkit:PlayerHeads-craftbukkit-support:jar:</include>
-                  <include>org.shininet.bukkit:PlayerHeads-legacy-support:jar:</include>
-                  <include>org.shininet.bukkit:PlayerHeads-craftbukkit-1.16-support:jar:</include>
-                  <include>org.shininet.bukkit:PlayerHeads-craftbukkit-1.13-support:jar:</include>
-                  <include>org.shininet.bukkit:PlayerHeads-craftbukkit-1.8-support:jar:</include>
-                  <include>org.shininet.bukkit:PlayerHeads-glowstone-1.12-support:jar:</include>
                 </includes>
               </artifactSet>
               <filters>
                 <filter>
-                  <artifact>org.shininet.bukkit:PlayerHeads-core:jar:</artifact>
+                  <artifact>org.shininet.bukkit:PlayerHeads-compatibility-api:jar:</artifact>
                   <excludes>
                     <exclude>com/github/crashdemons/playerheads/compatibility/CompatibilitySupport*</exclude>
+                  </excludes>
+                </filter>
+                <filter>
+                  <artifact>org.shininet.bukkit:PlayerHeads-compatibility-lib:jar:</artifact>
+                  <excludes>
+                    <exclude>org/shininet/bukkit/playerheads/events/*Event*</exclude>
                   </excludes>
                 </filter>
               </filters>
@@ -247,8 +248,24 @@
       <scope>test</scope>
       <exclusions>
         <exclusion>
+          <artifactId>commons-lang</artifactId>
+          <groupId>commons-lang</groupId>
+        </exclusion>
+        <exclusion>
+          <artifactId>guava</artifactId>
+          <groupId>com.google.guava</groupId>
+        </exclusion>
+        <exclusion>
+          <artifactId>gson</artifactId>
+          <groupId>com.google.code.gson</groupId>
+        </exclusion>
+        <exclusion>
           <artifactId>bungeecord-chat</artifactId>
           <groupId>net.md-5</groupId>
+        </exclusion>
+        <exclusion>
+          <artifactId>snakeyaml</artifactId>
+          <groupId>org.yaml</groupId>
         </exclusion>
       </exclusions>
     </dependency>
